art_snob_api/requirements.txt
aiohttp==3.6.2
appnope==0.1.0
async-timeout==3.0.1
attrs==20.1.0
backcall==0.2.0
backoff==1.10.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
decorator==4.4.2
ecdsa==0.14.1
fastapi==0.61.0
fastapi-cloudauth==0.2.0
future==0.18.2
gcloud-aio-auth==3.4.3
gcloud-aio-datastore==6.1.0
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-datastore==1.15.0
google-cloud-logging==1.15.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
jedi==0.17.2
joblib==0.17.0
multidict==4.7.6
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
requests==2.24.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sendgrid==6.5.0
starlette==0.13.6
threadpoolctl==2.1.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
yarl==1.5.1